Minor refactor for sql node convert logic
